pyproject.toml
[tool.poetry]
name = "zomantic"
version = "0.1.0"
description = "Semantic Scholar's recommentations from Zotero items"
authors = ["Ana Paula Gomes"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
peneira = "zomantic.main:main"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
